+++ swap/src/monero/wallet_rpc.rs
@@ -49,30 +49,30 @@ static MONERO_DAEMONS: Lazy<[MoneroDaemon; 12]> = Lazy::new(|| {
 compile_error!("unsupported operating system");
 
 #[cfg(all(target_os = "macos", target_arch = "x86_64"))]
-const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-mac-x64-v0.18.3.1.tar.bz2";
+const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-mac-x64-v0.18.4.0.tar.bz2";
 #[cfg(all(target_os = "macos", target_arch = "x86_64"))]
-const DOWNLOAD_HASH: &str = "7f8bd9364ef16482b418aa802a65be0e4cc660c794bb5d77b2d17bc84427883a";
+const DOWNLOAD_HASH: &str = "c35a4065147f8eeaa130a219e12e450fb55561efe79ded7d935fbfe5f7ba324c";
 
 #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
-const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-mac-armv8-v0.18.3.1.tar.bz2";
+const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-mac-armv8-v0.18.4.0.tar.bz2";
 #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
-const DOWNLOAD_HASH: &str = "915288b023cb5811e626e10052adc6ac5323dd283c5a25b91059b0fb86a21fb6";
+const DOWNLOAD_HASH: &str = "9d36ec8a1da1f31d654a8fd8527f4cae03545d8292bb1a2fe434ca454b3c0976";
 
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
-const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-linux-x64-v0.18.3.1.tar.bz2";
+const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-linux-x64-v0.18.4.0.tar.bz2";
 #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
-const DOWNLOAD_HASH: &str = "23af572fdfe3459b9ab97e2e9aa7e3c11021c955d6064b801a27d7e8c21ae09d";
+const DOWNLOAD_HASH: &str = "16cb74c899922887827845a41d37c7f3121462792a540843f2fcabcc1603993f";
 
 #[cfg(all(target_os = "linux", target_arch = "arm"))]
 const DOWNLOAD_URL: &str =
-    "https://downloads.getmonero.org/cli/monero-linux-armv7-v0.18.3.1.tar.bz2";
+    "https://downloads.getmonero.org/cli/monero-linux-armv7-v0.18.4.0.tar.bz2";
 #[cfg(all(target_os = "linux", target_arch = "arm"))]
-const DOWNLOAD_HASH: &str = "2ea2c8898cbab88f49423f4f6c15f2a94046cb4bbe827493dd061edc0fd5f1ca";
+const DOWNLOAD_HASH: &str = "b35b5e8d27d799cea6cf3ff539a672125292784739db41181b92a9c73e1c325b";
 
 #[cfg(target_os = "windows")]
-const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-win-x64-v0.18.3.1.zip";
+const DOWNLOAD_URL: &str = "https://downloads.getmonero.org/cli/monero-win-x64-v0.18.4.0.zip";
 #[cfg(target_os = "windows")]
-const DOWNLOAD_HASH: &str = "35dcc4bee4caad3442659d37837e0119e4649a77f2e3b5e80dd6d9b8fc4fb6ad";
+const DOWNLOAD_HASH: &str = "00151a96e96ef69eedf117c4900e6d0717ca074a61918cd94a55ed587544406b";
 
 #[cfg(any(target_os = "macos", target_os = "linux"))]
 const PACKED_FILE: &str = "monero-wallet-rpc";
@@ -80,7 +80,7 @@ const PACKED_FILE: &str = "monero-wallet-rpc";
 #[cfg(target_os = "windows")]
 const PACKED_FILE: &str = "monero-wallet-rpc.exe";
 
-const WALLET_RPC_VERSION: &str = "v0.18.3.1";
+const WALLET_RPC_VERSION: &str = "v0.18.4.0";
 
 #[derive(Debug, Clone, Copy, thiserror::Error)]
 #[error("monero wallet rpc executable not found in downloaded archive")]